backend/requirements.txt
flask
flask-cors
paho-mqtt
flask-socketio

